feasibility/requirements.txt
streamlit
openai
python-dotenv
ipykernel
matplotlib
